Bump for release v1.4.1.
Not sure what weblate will do if I copy an initial changelog based on the v1.4.0 release.
Will this come up as 'No translation requried'? Or will it say 'Incorrect translation' or something?
Hopefully this doesn't cause more issues than it tries to solve. The idea is to give translators
a bunch of context so they don't need to re-translate the changelog from a previous release.